changed long to int in the tag struct, as longs are 8 bytes in 64-bit architectures.
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@14261 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/mrpeach/net/udpsend~.h
@@ -84,12 +84,12 @@ typedef union _flint
 
 typedef struct _tag
 {                           /* size (bytes) */
-    char tag[4];            /*  4  */ /*"TAG!"*/
-    char format;            /*  1  */
-    long count;             /*  4  */
-    char channels;          /*  1  */
-    long framesize;         /*  4  */
-    char reserved[2];       /*  2  */ /* pad to 16 bytes */
+    char    tag[4];         /*  4  */ /*"TAG!"*/
+    char    format;         /*  1  */
+    int     count;          /*  4  */
+    char    channels;       /*  1  */
+    int     framesize;      /*  4  */
+    char    reserved[2];    /*  2  */ /* pad to 16 bytes */
 } t_tag;                    /*-----*/
                             /* 16  */
